+++ fastai/train.py
@@ -18,11 +18,11 @@ def fit_one_cycle(learn:Learner, cyc_len:int, max_lr:Union[Floats,slice]=default
                                         pct_start=pct_start, **kwargs))
     learn.fit(cyc_len, max_lr, wd=wd, callbacks=callbacks)
 
-def lr_find(learn:Learner, start_lr:Floats=1e-7, end_lr:Floats=10, num_it:int=100, **kwargs:Any):
-    "Explore lr from `start_lr` to `end_lr` over `num_it` iterations in `learn`."
+def lr_find(learn:Learner, start_lr:Floats=1e-7, end_lr:Floats=10, num_it:int=100, stop_div:bool=True, **kwargs:Any):
+    "Explore lr from `start_lr` to `end_lr` over `num_it` iterations in `learn`. If `stop_div`, stops when loss explodes."
     start_lr = np.array(start_lr) if is_listy(start_lr) else start_lr
     end_lr = np.array(end_lr) if is_listy(end_lr) else end_lr
-    cb = LRFinder(learn, start_lr, end_lr, num_it)
+    cb = LRFinder(learn, start_lr, end_lr, num_it, stop_div)
     a = int(np.ceil(num_it/len(learn.data.train_dl)))
     learn.fit(a, start_lr, callbacks=[cb], **kwargs)